+++ src/modules/esm.mjs
@@ -1,0 +1,47 @@
+import path from 'path';
+import { release, version } from'os';
+import { createServer as createServerHttp } from 'http';
+import './files/c.js';
+import a from './files/a.json' assert {type: 'json'};
+import b from './files/b.json' assert {type: 'json'};
+import {fileURLToPath} from "url";
+
+
+const currentFolder = path.dirname(fileURLToPath(import.meta.url));
+const currentFile = fileURLToPath(import.meta.url);
+
+const random = Math.random();
+
+let unknownObject;
+
+if (random > 0.5) {
+    unknownObject = a;
+} else {
+    unknownObject = b;
+}
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${currentFile}`);
+console.log(`Path to current directory is ${currentFolder}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export {
+    unknownObject,
+    myServer,
+};
+
